Create TODO in BuildOpenCV.cmake Create include directory Refactor directories Add include_directories to main CMakeLists.txt
