+++ external/opencv/CMakeLists.txt
